ui updates and folder restructuring
