+++ src/wedding-admin-cms/ClientApp/src/App.jsx
@@ -1,18 +1,19 @@
 import React from 'react';
-import { Route } from 'react-router-dom';
+import { Route, Switch } from 'react-router-dom';
 import { Layout } from './components/Layout';
 import Home from './pages/Home';
-import FetchData from './pages/FetchData';
-import Counter from './pages/Counter';
+import EditWedding from './pages/EditWedding';
 
 import './custom.css'
 
 const App = () => {
   return (
     <Layout>
-      <Route exact path='/' component={Home} />
-      <Route path='/counter' component={Counter} />
-      <Route path='/fetch-data' component={FetchData} />
+      <Switch>
+        <Route exact path='/' component={Home} />
+        <Route path='/edit-wedding' component={EditWedding} />
+        <Route path="*" component={() => <div>Page not found! Try again!</div>} />
+      </Switch>
     </Layout>
   );
 };
