+++ src/wedding-admin-cms/ClientApp/src/components/NavMenu.jsx
@@ -22,10 +22,7 @@ export const NavMenu = () => {
                 <NavLink tag={Link} className="text-dark" to="/">Home</NavLink>
               </NavItem>
               <NavItem>
-                <NavLink tag={Link} className="text-dark" to="/counter">Counter</NavLink>
-              </NavItem>
-              <NavItem>
-                <NavLink tag={Link} className="text-dark" to="/fetch-data">Fetch data</NavLink>
+                <NavLink tag={Link} className="text-dark" to="/edit-wedding">Edit Wedding(s)</NavLink>
               </NavItem>
             </ul>
           </Collapse>
